Add a trunk stable flag for reordering wallpapers during a user switch
If the target user's home and lock screen wallpapers are different,
both of them are loaded during the user switch and this lengthens
the user switch duration unnecessarily. We can load the necessary one
during the user switch and postpone loading the other after the switch
depending on whether the target user has credentials or not.

Bug: 324911115
Test: m all_aconfig_declarations
Flag: ACONFIG android.multiuser.reorder_wallpaper_during_user_switch DEVELOPMENT
Change-Id: Ia5272cedc8bb1c92ec76f5854b7c697a416e02e5