SharedMemory: mark fdsan ownership
We noticed some internal test failures that were due to fdsan aborts
from SharedMemory.Closer getting its fd stomped on and then aborting
when it tried to close an fd owned by a unique_fd that got allocated in
its spot.

This immediately found a bug in our code where we were creating a
ParcelFileDescriptor from SharedMemory.getFileDescriptor, along with a
bug in ContentResolverTest.

Test: `atest android.os.cts.SharedMemoryTest android.content.ContentResolverTest`
Change-Id: I3fd4a5c5e478e0674021ffed3d624046e978dfdc