Set resolvedType when starting activity in TaskFragment
Obtain and set the `resolvedType` of intent to the
`ActivityStarter.Request`.

Prior to the change, certain intents were incorrectly resolved due to
the lack of the `resolvedType` field. An example intent that was
incorrectly resolved prior to the change:

Intent(Intent.ACTION_VIEW, Uri.parse("content://com.android.calendar/time/<unix_time>"))

Also adds a regression test associated with this issue.

Bug: 326133028
Test: atest WmTests:ActivityStartControllerTests
Change-Id: I9c126d29e15c252ee921edfa9858ae4ec6df5558