Improve multi-crop forward compatibility logic
Make sure we first adapt the existing crop for the new device before
generating missing crops. This is important if there is a B&R from a
similar device with slightly different screen dimensions: we don't want
to calculate the missing crops based on the crops of a previous device;
we first want to adjust the existing crops for the new device before
calculating other crops for missing orientations.

Flag: aconfig com.android.window.flags.multi_crop DEVELOPMENT
Bug: NA (let's not spam bugs with small fixes)
Test: atest WallpaperManagerTest WallpaperCropperTest
Change-Id: I199d6b2c00314c00ca6c240b942e5b0c4e902be9