Fix more options crash
Before this fix, the pending intent was created in the Systenm
Server and a fillInIntent was created in the CredentialAutofillService.
This intent carried icons which could be of type bitmap too. This caused
a crash when invoking startIntentSender from the AutofillManager
because it is not possible to crete an intent with a bitmap in one process
and invoke it from another (cannot write file descriptor).
With this fix, the System Server only passes an intent with the request
Info. A pending intent per autofillId is genrated in CredentialAutofillService
and the underlying intent with which the pending intent is created contains
the icons.

Test: Cts
Bug: 325008438

Change-Id: Ic46984fed6cea0fd2e4af69e5d6dee82cdfc1d20