Primary screen display optimization for the hero use cases
When there's a single account, the primary screen will remove credential type name (for password & passkey entries only) and credential provider  name (for all types of entries) since the removed info will already be reflected from the title and the feature mentioned in the next paragraph.

Additionally, when all the accounts on the primary screen come from the
same provider, we will display that provider name & branding on top of
the title.

Bug: 326493264
Test: screenshots attached to bug
Change-Id: If87f8fe67e64f237d97591d24762b937080046b7