Fix vibration finished before HAL callback
Update TurnOffVibratorStep so it will be considered part of the main
vibration steps when it's waiting for the HAL completion callback.

This keeps the ramp down steps outside the main vibration duration,
but it makes sure the vibration will be completed after a callback is
received, even if delayed compared to the expected duration.

Also fix the VibratorManagerService command line to wait for the
vibration ramp down and clean up before ending the command, to prevent
vibrations from being cancelled because of binder death.

Fix: 327608323
Test: atest VibrationThreadTest
Change-Id: I9cf0e1263f73b719de16782a85df87f740c9c5d0