Clear additional subtypes upon background APK updates
Previously

  MyPackageMonitor#onFinishPackageChangesInternal()

took effect when and only when the APK change was for the current
user, which means that additional subtypes would remain unchanged if
the APK update happened while the user is not
InputMethodManagerService's current IME user.

With this commit, additional subtypes start being cleared even for a
background APK update.

See the corresponding CTS change [1] about detailed behavior change.

 [1]: Iecce1e7c035bb20961b3ac68414f212c4b748da6

Bug: 322062773
Fix: 27859687
Test: atest CtsInputMethodInstallTestCases:AdditionalSubtypeLifecycleTest
Change-Id: I10aa547f0de607ef3c0ba26764dac0585c40c843