DMD fix flicker vote
In ag/25829274, we made some code in DMD work for multiple displays, so what happened with mRefreshRateChangeable was that it was being set based on the min and max refresh rates of the display that was most recently updated. When using Screen Recording, a new virtual display is added with only one refresh rate 60, so mRefreshRateChangeable was being set to false, even though there are other displays with changeable refresh rates.

Bug: 25829274
Test: adb shell dumpsys display - look at mRefreshRateChangeable and PRIORITY_FLICKER_REFRESH_RATE
Test: atest com.android.server.display.mode
Change-Id: Ic05ab0401766f6d0edd295ca337b0713f7e367a5