Use detachFd for getPlanes in CameraExtensions
Fence of ParcelImage is ParcelFileDescriptor, when call getFd() to
get file descriptor of it, the file descriptor still be owned by
ParcelFileDescriptor, which cause fatal occurs when close the
file descriptor in Gralloc4Mapper.

Therefore, use detachFd() to get file descriptor of ParcelImage Fence.
When call detachFd() to get file descriptor of ParcelFileDescriptor,
the file descriptor will not be owned by ParcelFileDescriptor,
so the file descriptor can be closed normally in Gralloc4Mapper.

Bug: 334802005
Change-Id: I8388c913f967565d90eca80e6b1b8ccb09ab018e