Replace IMMS#mSettings with IMMS#mCurrentUserId (2nd try)
This is the 2nd attempt of submitting my previous CLs [1][2], which got
reverted due to a boottime regression (Bug 329414077) [3][4]. We have
already addressed the major contributor of the boot time regression [5],
so this CL should be ready to be submitted again.

This CL removes

  InputMethodManagerService#mSettings

as the same instance can be obtained from

  InputMethodSettingsRepository

as long as we keep track of the current user ID.

This is a mechanical refactoring CL. There must be no user observable
behavior change.

 [1]: Idae0fb6d1a9182efe917f81a1d881d488a1b57e5
      7c9ede18b7396ddef4786e53ce6d2a9de22177bd
 [2]: I4b7dc1b0e52e9bf25095eb2e3cc520d46b3869e5
      ccf0ae58a50d7d08fa388925e0cc3c447ad5fca2
 [3]: I0d65cb9edcdbdaea45b1fa80abfe917b72fb9a50
      d2b84e1bee7d6a4273a5252f136ba2339b04eb46
 [4]: I585e5c6745c40e51f1080b78367175990b2c5884
      2350847c8a54c3b93d9a03969e59cc749ae89fc1
 [5]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      c133d065801ed9c77b20112c25d872472b34e833

Bug: 305849394
Fix: 329163064
Fix: 333591545
Test: presubmit
Change-Id: Iaa7f08c9081aac9abe0c95d522185632d575d6b8