Play cross-activity back animation without letterboxes
When the closing and entering activity are both letterboxed and they have the same localBounds, let's exclude the letterbox areas from the animation and play the animation on the actual activity content instead. To do so, the letterbox areas are cropped from the animation targets and artificial letterbox layers are added instead.

Bug: 328446606
Flag: ACONFIG com.android.window.flags.predictive_back_system_anims NEXTFOOD
Test: Manual, i.e. testing back animation in letterboxed app (and verifying no regressions in landscape, vertical and horizontal split screen, large screen, light mode, dark mode, resizeableActivity=false etc.)
Change-Id: I838c1bbc41949a69214f170501c777f4f305703b