Move intermittent calculation to ViewRootImpl from View
Bug: 334006209

Intermittent calculations can be done on a per-ViewRootImpl
basis rather than per-View basis and still retain 95% of the
effect. This will save calculations on every drawn View and
make one calculation per ViewRootImpl instead.

This CL also makes a fast path for votePreferredFrameRate()
that handles when no values are set for velocity and
frame rate. This is the most common scenario and should
have a fast-path.

Test: ran performance test, ViewRootImplTest, ViewFrameRateTest
Change-Id: Ib103287a976b0b3fc7a246e5780f6d6a8be58133